Update the Maven Central publishing URL to avoid sunset service (#3457)
The service we were using to publish to Maven Central was sunset. There
is a legacy API that allows us to continue to use our old publising
gradle platform, and so I've updated the configuration so that it
follows the advice available on their website:
https://github.com/gradle-nexus/publish-plugin?tab=readme-ov-file#publishing-to-maven-central-via-sonatype-central

This is advertised to be pretty straightforward. The two main steps are
(1) update the publishing URL to one that uses a bridge API and (2)
modify the credentials to ones generated by the new portal. I've done
that (and already updated the secrets on the repo with the new
credentials). So we should be able to get a new build published through
the new gateway once this is in.

I've also gone ahead and removed the release notes for the builds that
never actually were published. That way, when we get a new build out,
we'll get the correct set of release notes automatically without needing
to tweaks those by hand later.

This fixes https://github.com/FoundationDB/fdb-record-layer/issues/3456.