Escape the ' character in the loggable method to help with log parser frameworks. (#3460)
The ' (single quote) character can confuse log parsers. This PR escapes
this character when used within a log message.
This change alters the serialized representation of the byte array as a
string. Any usage of this representation would be therefore modified.
Searching the codebase showed usage within Exception messages and
logging, which is likely OK, but care should be taken to verify that no
ill impact is exposed.
Resolves #3461