Add support for NULL type fields in ExecuteInsert (#3490)
This PR adds support for inserting structs with NULL typed fields into
the table via `ExecuteInsert` in the Direct Access API. Also adds a test
for the same.

This bug was introduced by
https://github.com/FoundationDB/fdb-record-layer/pull/3484, which
essentially shifts the Relational-layer from using Sql Type codes for
internal operations to relying on Relational DataTypes for getting a
richer information.

Fixes: https://github.com/FoundationDB/fdb-record-layer/issues/3491