Add test to verify that temp functions are created in TxnBound database (#3488)
This PR replaces the `NoOpMetadataOperationsFactory` passed in
`TransactionBoundSchemaTemplate` with an implementation of
`MetadataOperationsFactory` that supports
`getCreateFunctionConstantAction` in the same way as it is supported by
the `RecordLayerMetadataOperationsFactory`.

Also adds the test to verify that the TEMPORARY FUNCTION is actually
available in the txn-bound schema template in the case when the
`Transaction` is externally provided to the connection.

fixes https://github.com/FoundationDB/fdb-record-layer/issues/3489

---------

Co-authored-by: Youssef Hatem <y_hatem@apple.com>