Initial support for DROP TEMPORARY FUNCTION (#3492)
This PR provides an initial support for ‘DROP TEMPORARY FUNCTION’ DDL
statement. It does so by creating a new ConstantAction-based
ProceduralPlan that does the simple remove key operation from the
invokedRoutines in transaction-based SchemaTemplate.

Note that it does not evaluate if other functions transitively depend on
the function marked for deletion and hence, does not perform cascading
delete or rejects the current delete. As a result, compiled dependant
functions still succeeds. This could be solved by having some dependency
graph in the schema template. This is out of the scope of the current
PR. https://github.com/FoundationDB/fdb-record-layer/issues/3493

Also includes tests for the same. This PR also adapts
TransactionBoundDatabase to provide drop temporary function operation to
it as well.