Support unsetting transaction-bound SchemaTemplate in a Transaction (#3496)
This PR add support for "unsetting" a transaction-bound SchemaTemplate
from the ongoing transaction. This is essential if we want to revert the
execution to using the SchemaTemplate from the catalog thereby undoing
all the temporary function definitions at once. Note that this could be
more efficient than the `DROP TEMPORARY FUNCTION ...` DDL introduced in
the previous PR and have the wider scope that is beyond `TEMPORARY
FUNCTIONS` definitions.

Also added a test for testing it in the context of
`TransactionBoundDatabase`.