Omit portions of stack trace that are just junit
I ran two tests that had exceptions deep in lucene code, with suppressions
and causes, and this omission reduced the 768 lines / 76,114 bytes
to 183 lines / 16,946 bytes.
GitHub has a limit of 1024k for the summary, so this should help
keep us below the limit, and make it easier for developers to
diagnose why a test failed.