Introduce option to setup transaction before executing queries (#3471)
This adds the ability to specific a transaction setup to yamsql.
More specifically for an individual query you can have: `setup:` which
will be executed at the start of every transaction for that query

To avoid writing the same setup repeatedly, it also adds a new block:
`transaction_setups:` which allows you to create references for the
setups. Inside the query, it can now be referenced via the key, using
`setupReference:`.

This can be useful if you want to test with temporary functions, but
still have all the benefits of randomized, and forced-continuations.

This currently only supports a single query in the setup, but it
wouldn't be hard to extend it to take an array.