Introduce new timers for the different phases of the Cascades planner
This commit introduces new StoreTimer.Events that record how long each phase (rewriting, planning)
of the Cascades planner take. To achieve this, the Relational layer passes the instance of StoreTimer
within the current transaction to the Cascades planner upon creation, which it makes use of by recording
the time a phase took using a new Cascades task, `FinalizePlannerPhase`, which simply calls record on the
StoreTimer instance (if it is not `null`) with the total time the planner phase took. The
`FinalizePlannerPhase` task is schedueled to be executed by the `InitiatePlannerPhase` task, ensuring that
it is the last task that is run for that phase.

In addition, the logs for queries (which happen if debug logging is enabled or if a query is too slow) is enhanced
to include the durations for the two planner phases.