Cleanup 4.5.3.0 Release Notes
We ended up pushing the 4.5.2.0 and 4.5.3.0 releases in quick succession because the publishing for 4.5.2.0 was disrupted. We were able to publish that release anyway, so in theory we could leave the 4.5.2.0 release out, but it seems more consistent with our other practices to just clean up the release notes to remove the 4.5.2.0 reference.