Remove synchronized blocks from FDBDirectoryWrapper (#3494)
ForkJoinPool doesn't play well with `synchronized` blocks, particularly,
it can result in thread starvation / deadlocks. Threads waiting on
`synchronized` blocks, still count as active, and thus count towards the
parallelism of the pool.
The tests here run in a very small pool, to make sure that this won't
happen, but in actual production workloads, one would expect a much
larger pool, but also a lot more concurrent activity.

These tests were failing pretty reliably before with a pool size of 3,
and now they seem to pass reliably.

One change in here that may be surprising is the change around the
`writerReader`. I looked at the usages, and there were two:
1. Cursors would call `getReader` which would always flush, and thus
never use the field
2. `LuceneIndexMaintainer.deleteDocument` would never flush

So it makes sense to have these be two independent calls, which allows
`getWriterReader` to use a `LazyCloseable`.

Resolves: #2989
Resolves: #3146