Upload coverage data to teamscale (#3519)
After a Pull Request workflow, this will be queued to download the
coverage data, and upload it to teamscale.

This has to be done as a separate flow from the actual `Pull Request`
workflow, because `pull_request` triggers do not have access to secrets,
since they run code from the open pull request.

This does take advantage of this action:
https://github.com/cqse/teamscale-upload-action

If the `Pull Request` workflow fails, this will be skipped.