Split the Cascades Debugger interface around stats and symbols
This commit splits the com.apple.foundationdb.record.query.plan.cascades.debug.Debugger interface into two
separate interfaces (StatsDebugger and SymbolDebugger) which both extend the base Debugger interface, in
order to ensure that operations related to recording Debugger events (with the purpose of calculating stats
based on the profiling of these events) make use of the StatsDebugger interface, and operations related to
keeping a registry of all Symbols used within the planner for purposes of associating them with friendly names
make use of the SymbolDebugger interface.

In addition to splitting the Debugger interface, the State class is also split into two classes, one that keeps
of the Debugger events and stats around them (EventState) and one that keeps track of the current Symbols seen
during planning (SymbolTables).