Group Debugger events by PlannerPhase in stats
This commit modifies the EventState class and the StatsMaps that it produces
to group Debugger events by the planner phase it was emitted during to enable
the ability of getting stats for each planner phase separately.

Given that some Debugger events (e.g. InsertIntoMemoEvent) do not store the
planner phase they were emitted in (as it not visible to the code that emits
them), these events are stored separately and can be retrieved from the
StatsMaps (in addition to the ability of retrieving all events, regardless
of the planner phase that they were emitted in).