Add new lightweight debugger implementation
This commit adds a new implmentation of the Cascades Debugger interface,
which only keeps track of statstics around debugger events and doesn't support
keeping track of symbol tables. This new implementation should be safe to
use outside of tests since it should have a minimal impact on the planner's
performance.