Handle metrics for the same query but with different setups (#3531)
Planner metrics previously assumed that it was only reasonable to have a
query once within a single block. Having setup can make this more
complicated, so this changes the keys for the metrics to have the setup
too.

There were a couple unrelated cleanups along the way:
1. `markDirty()` was annotated `@Nullable`, but is `void`
2. `putMetrics` took `isDirectyMetrics`, but that parameter was never
used, so I removed it.