Add serialized plan complexity to the continuation info in EXPLAIN
This commit adds a new column, 'SERIALIZED_PLAN_COMPLEXITY' to the
'PLAN_CONTINUATION' struct in the result from EXPLAIN which contains the complexity
of the plan serialized within the continuation, if the continuation contains a compiled
statement. This new column is populated for EXPLAINs on an 'EXECUTE CONTINUATION' statement
and a regular SELECT statement that uses the `WITH CONTINUATION` clause.