Add a metric to measure generate continued plan (#3553)
This PR adds a new metric to measure stats about obtaining a physical
plan from the continuation. This pathway is not currently accounted for
and hence, we do not have visibility of the time spent in "planning" a
query but not in "looking up in cache". `EXECUTE CONTINUATION ...`
follows this code path and accounts for all the requests that register
the `GET_PLAN` event but does not register `CACHE_LOOKUP` event.