Add new wrapper value for removing duplicates from array values
This adds a new value, ConstantArrayDistinctValue, which can be used to
wrap an underlying constant array value to remove duplicates from the
underlying array during evaluation.

The purpose of this new value is to use it to wrap in-lists used in
InJoin, InUnion and FlatMap plans to avoid duplicate records if the
in-list contains duplicate values.