Verify continuation plan hash against serialized plan hash
This commit asserts that the plan hash in the continuation passed to
ContinuedPhysicalQueryPlan::getSerializedPlanFromContinuation matches
the plan hash of the plan that was extracted from the continuation and
used to construct the ContinuedPhysicalQueryPlan.

In practice, this should always be true unless the code was changed in
an incompatible way so the continuation passed to getSerializedPlanFromContinuation
is not the same continuation that was used to deserialize the physical
plan.