When correcting yamsql metrics, only update queries with differences
Without this change, if you add a new query and run the test with
CORRECT_METRICS, it will update all the timings for all of the
other queries. This is less important for the binp, but can make it
hard to look at the metrics yaml diff, as there will be many lines
to ignore.

With this change, it will keep the original info for every query
that did not result in a diffence.