Ensure that `TransactionalRunner::runAsync` closes a context that it creates even if future creation fails (#3562)
There was a case that could result in a transaction being created that
was not closed. One thing to note is that the exception actually is
closed with the current set up when the `TransactionalRunner` is closed,
so this leak should only be a problem if the runner isn't being closed
(an API error), but it would be good anyway to ensure that this gets
closed.

This fixes #3561.