Instead of adding coverage to release add all tests to nightly
We may want to make it only run the test/mixed-mode-test if there
have been changes, which is do-able, I think by checking the git
log to see if anything was merged in the last 24 hours. But that
doesn't work super great with the fact that we don't run nightly
over the weekend. But we still want to run the nightly tests if
there were no changes, as it adds randomness.