Start two fdb clusters in CI
This took a lot of trial and error in CI, and adding and
removing a lot of logging and investigative code.
The key thing that tripped me up was the fact that
you have to do `configure new ssd` if it is a new cluster,
which we were not doing previously, because the
installation created the cluster.