Composite aggregates (#3266)
This PR adds the ability to plan arbitrary aggregation queries using
aggregation indexes as well as intersections of aggregation indexes. It
also corrects/adds the infrastructure to deal with complicated order-by
requirements.

- refactor data access rules into a data access rule for value indexes
and one for aggregation indexes
- have both of these data access rule implementations share the maximum
code (only intersection planning is different)
- intersection planning for aggregation indexes
- because of the same reasoning as regular index intersections AND in
order to answer the query as these indexes carry data that is not part
of the underlying table (the aggregation)
- rollups of finer-granularity index scans if such a roll-up can be
useful in answering the query
- order by requirements are treated in the same framework as all other
kinds of expressions (using M3 maps, translations, etc.)