Upload coverage from nightly builds (#3568)
Add new jobs to the nightly flow to run standard tests and mixed-mode
tests. This is primarily there so that we can upload coverage to
teamscale.
These new teamscale steps are set to continue-on-error, to make sure
they are working correctly before impacting the nightly build.

I also extracted the teamscale upload into a separate action to decrease
repetition, this is the biggest risk; if I messed this up, it could
impact the post-pr teamscale upload.