When correcting yamsql metrics, only update queries with differences (#3563)
Without this change, if you add a new query and run the test with
CORRECT_METRICS, it will update all the timings for all of the other
queries. This is less important for the binp, but can make it hard to
look at the metrics yaml diff, as there will be many lines to ignore.

With this change, it will keep the original info for every query that
did not result in a diffence.

Here I have a demo of how this looks with this change:
https://github.com/ScottDugas/fdb-record-layer/commit/b32570a62c70b5608ffb97e7091faad7919a3c0e
And here is how it looks if I run it without this change:
https://github.com/ScottDugas/fdb-record-layer/commit/8f0778f