Support SQL array subscript operator (#3586)
This introduces support to SQL array subscription operator, with the
following caveats:
- It works with one-dimensional arrays only.
- Similar to other vendors, such as Postgres and Oracle, it returns
`null` when the index is out-of-bound instead of raising an error.
- Uses 1-based index, as defined in SQL standard.

Example:

```sql
create table A(pk integer, x integer array, primary key(pk))
insert into A values (1, [1, 2, 3]), (2, [2, 3, 4]), (3, [3, 4, 5])
select X[2] from A where pk = 1;
-- returns 2
select X[1+1] from A where pk = 1;
-- returns 2
select X[-1] from A where pk = 1;
-- returns null
select X[1000] from A where pk = 1;
-- returns null
```

This also restructures the parsing rules for expressions such that it
prohibits more invalid combinations of binary expressions, and aligns
them to work well with the subscript operator, and adds support to `not
in` which was previously throwing a syntax error.

This fixes #3584, #3570, and partially #3588 with few limitations.