Make YAML tests (and SQL more generally) able to encrypt records at rest (#3557)
First change handling of `Options` to be able to pass them around via
gRPC to JDBC connections. (These were formerly in #3569.)

Then use options to make the `StoreConfig` and have that _create_ the
`Serializer`, rather than receiving it from some earlier config. New
encryption options make use a key store file to hold the secret key.
Such a file is checked into the YAML test directory with a couple of
keys. Hopefully this will make it clear that these keys are for the YAML
tests only and production deployment needs something more sophisticated.