Set codeCoverageData so that you _can_ run it with tests
The usage of `fileTree` caused codeCoverageReport to have implicit
dependences on every task that ran, but this is not what we want.
I updated it to more directly depend on the jacoco output.
I also updated it to explicitly mustRunAfter all test and jar
tasks of the relevant sub projects.
By using mustRunAfter instead of dependsOn, I think it should
allow running codeCoverageReport without running jar or test,
which will work better with the pull_request workflow, where
we run the tests in different jobs, and bring in the results
to run codeCoverageReport.