Remove duplicates from IN-list before turning it to an ExplodeExpression
This commit makes use of the new wrapper value introduced in 4.5.10.0,
ArrayDistinctValue, to remove duplicate values from the list provided to
an IN predicate in SQL before transforming the IN predicate to an
ExplodeExpression. This prevents duplicate result values if the planner
ends up implementing the query graph as a nested loop join between the
ExplodeExpression of the IN list and the left operand of the IN
predicate.

This is a breaking change that is not compatible with versions of
fdb-record-layer < 4.5.10.0.