Fixup RankIndexTest.writeOnlyRankQuery
The reason for the issue is that it was reassigning
`fdb = dbExtension.getDatabase()`, but it loaded data into `fdb` in
a `beforeEach`, so if it picked different clusters it wouldn't have
any data, and execute wouldn't fail.
In debugging I also removed teh second `openRecordStore`, because
the state should respond to calling `markIndexWriteOnly`, you shouldn't
have to re-read, and we actually say you should never open the same
store with different objects in the same transaction.

I also isolated the exception check, and switched to use assertJ for
the assertion, because it ends up being quite a bit cleaner than with
junit's assertions.