Run jar when calculating codeCoverageReport in nightly build (#3613)
Some projects were showing up, incorrectly with 0% coverage. The
resolution for this in PRB was to run `jar` and pass around the
artifacts. We don't have to pass around the artifacts, so we can just
run jar in the same command. I'm hoping that the change here also fixes
the issue with the nightly build.

I validated locally running `jar test destructiveTest
codeCoverageReport`, and the `mixedMode` version. It does not complain
about implicit task dependencies.