Extend serialized record encoding to allow for multiple encryption keys (#3522)
A key number is encoded in the serialized record, in a format that is
backward compatible. An interface can be provided to resolve key numbers
into keys for encryption and decryption. The byte format of existing
encrypted data is such that it is interpreted as having key number 0.

Compatibility of this change in persistent data is checked by the
`encrypted.yamsql` test when run against `4.6.1.0`. The YAML framework
still uses the old serializer API to set encryption keys, so this also
demonstrates that this is compatible.

A followup PR will change YAML to use the new serializer API while
continuing to pass mixed tests.

---------

Co-authored-by: Scott Dugas <scott.dugas@gmail.com>