Create a new fdb-test-utils sub-project for some test utilities
This moves things that were in `fdb-extensions` and needed for testing
out into their own test-specific sub-project. The new project doesn't
depend on any of the other sub-projects.

I left `MultipleTransactions` behind in the testFixtures. It currently
depends on `FDBError` (which is part of fdb-extensions), and currently
no other project is depending on it, so I felt that leaving it
made sense.