Make use of key store its own key manager (#3606)
Following on from #3522, extend how key store is used by YAML tests to
allow for multiple keys in the connection options. Any one of these keys
can be read back.

The existing YAML tests continue to work with this new implementation,
as the options are known to the old serializer implementation and the
earlier PR's implementation of a single key.

A new YAML test shows the use of multiple keys, restricted to the
current version, as it involves a new option.

---------

Co-authored-by: Scott Dugas <scott.dugas@gmail.com>