remove CONTINUATION_CONTAINS_COMPILED_STATEMENT usage (#3650)
This PR does a couple of things:
- removes the usage of `CONTINUATION_CONTAINS_COMPILED_STATEMENT` option
in Connection and as a query option.
- The option was, by default, set to true and is now true for all the
requests to the engine.
- We still support `SELECT ... WITH CONTINUATION ...`, which could be
removed separately after this work. Hence, it could happen that a query
started with a non-serialized continuation, but has now on serialised
continuation.
- JDBC driver still supports the option, for backward compatilibility.
However, the current change always sets that to `TRUE` and rejects
connections that have it set to `FALSE`. Once this is rolled out, we
should probably stop failing on false and mark the option as deprecated
in proto. Finally, we can stop setting the option and remove the option
altogether.
- Removes a few tests around testing this option with False.