Implementation of grouping size aggregator (#3577)
This implementation is creating a flavor of the
`SizeStatisticsCollectorCursor` that can return sizes aggregated by
sub-subspaces.
This relies on the original implementation but had to go into a separate
class as the generic type of the cursor has changed.
Eventually, the original cursor can be replaced by this one with the
aggregation size being "0".
For backwards compatibility, the original class was marked as
"deprecated". The original `SizeStatisticsResults` was marked as
deprecated and a new (non-embedded version) of it was added.

Initially, I thought that we can make a backwards-compatible change to
the `SizeStatisticsCollectorCursor` class, but that proved more
complicated with the changes to the result type, so I decided to go with
a new implementation instead.
 
Resolves #3576