Make YAML tests pass connection options to servers for transaction setups (#3638)
This PR makes transactional setups within a `test_block` in YAML tests
make use of the specified `connection_options` for a block when making
use of a gRPC server for testing, which is now possible thanks to the
work introduced in #3557 to pass the `connection_options` as part of the
request for the remote JDBC connection.

In addition, this PR also makes it possible to specify
`connection_options` for a `setup` block, making it possible to use
different connections options for different setup blocks within the same
YAML test.