Replace `String.format` calls with string concatenation where possible (#3660)
This goes through our codebase and audits our `String.format` usage. It
makes a few changes:

1. If the `String.format` call can be replaced with `String`
concatenation, it does so. This mainly means taking cases where strings
or numbers are naively placed together (e.g., `String.format("%s -> %d",
myStr, myInt)`) and replacing it with the equivalent concatentation
(e.g., `myStr + " -> " + myInt`). The reason this is done is efficiency:
it is more efficient to construct the strings and then concatenate them
than to use the Java `String` formatter most of the time, mainly because
the format string parsing is done at runtime with every invocation
rather than compiled into some kind of formatting object.
1. In places where that is more difficult, this ensures that the
`String.format` call has set `Locale.ROOT` to avoid localization
problems. This is what caused #3645, as the `MetricsDiffAnalyzer` was
formatting numbers with decimals, and then test asserts were firing if
the test ran with a different locale, which is a thing we do in our
nightly tests. In theory, that's actually sort of appropriate for a tool
like this (for example, it would make sense to localize `3.5` to `3,5`
if run on a system where that is the convention), but the rest of the
product doesn't support it, and it's not like the actual text is
localized.

When running on a system with `Locale.ROOT`, this should be a
transparent change.

This fixes #3645.