remove sql_type_code remove usage (#3651)
This PR is a follow-up PR in the series for deprecating using
sql_type_codes in JDBC, and instead using "richer" Relational types. The
support for relational types were introduced some time bakc in
https://github.com/FoundationDB/fdb-record-layer/pull/3510. This PR
shifts to relational type usage, away from sql_type usage, to make way
for retiring the latter field completely from the protobuf. After this,
we could remove the field in 2 steps:
1. stop setting jdbcSqlType fields in JDBC proto and mark them
deprecated.
2. remove from protobuf