Create a new fdb-test-utils sub-project for some test utilities (#3631)
This moves things that were in `fdb-extensions` and needed for testing
out into their own test-specific sub-project. The new project doesn't
depend on any of the other sub-projects.

I left `MultipleTransactions` behind in the testFixtures. It currently
depends on `FDBError` (which is part of fdb-extensions), and currently
no other project is depending on it, so I felt that leaving it made
sense.

This is in support of:
https://github.com/FoundationDB/fdb-record-layer/issues/3574 -- see
https://github.com/FoundationDB/fdb-record-layer/pull/3575#discussion_r2375582275