catch IndexOutOfBoundsException when accessing fields from metadata (#3672)
Accessing fields in struct metadata that are actually out of range, used
to throw the underlying exception without actually wrapping it into
SQLException, as required by the API. This was intrduced by a previous
PR that changed the struct and arrays metadata to be backed by
Relational's DataType rather than relying on sql type codes. This PR
fixes this.