Add support for testing against multiple FDB clusters (#3575)
This changes the way we specify cluster files in tests. It replaces the
`fdb-environment.properties` with `fdb-environment.yaml`. The new format
allows for specifying multiple cluster files.

The majority of tests now randomly chose a file at the start of the test
from the list of the available ones.
The following sub-projects set the environment variable in the gradle
task because getting the project under test to support custom cluster
files seemed like a bigger task:
- fdb-relational-cli
- fdb-relational-jdbc

Our CI builds also, now, setup two FDB clusters, and create an
`fdb-environment.properties`.

I added one new "test" to validate this all works. It connects to two
different clusters. It `assume`s that there are at least two clusters.
At some point in the future we will want to make that a stronger
contract to ensure that a bug in our CI setup doesn't cause us to not
run all the multi-cluster tests.

I created Draft PR:
https://github.com/FoundationDB/fdb-record-layer/pull/3623 to ensure
that everything still works if you don't specify `fdb-environment.yaml`
I also created this PR off the same version of main to confirm the test
counts didn't change:
https://github.com/FoundationDB/fdb-record-layer/pull/3621 Apparently we
do merge main to do the tests so PR #3621 saw fewer tests than this
branch when it was last run. I re-ran and they may be the same depending
on whether new tests were merged. I validated some of the differences,
and they look to be expected.

Resolves: #3574