Remove aspects of KeySpacePath for storing a resolved key
Shortly before ResolvedKeySpacePath was introduced, KeySpacePathImpl
would store the resolved tuple and remainder in itself.
The methods around this have been deprecated for years.
More pressingly, what having this value means is that if you have
a KeySpacePath that is generated from KeySpaceDirectorys and call
toTuple or toSubspace it will resolve in the context that you
provide, but if that KeySpacePath came from resolving a key then
toTuple / toSubspace will contain the original directory-layer
mappings, which could be problematic if the provided context
is not on the same cluster.