Add new KeySpacePath.exportAllData (#3566)
This introduces a new KeySpacePath.exportAllData to export all the data
stored in the path. This can eventually be used to import into another
cluster, or back into the same cluster, after clearing.

Other than the path information, the data exported is raw bytes, with no
transformation or indicated semantics.

Resolves:
[3572](https://github.com/FoundationDB/fdb-record-layer/issues/3572)