Protobuf compliant translation for supporting richer identifiers for tables and columns (#3696)
This PR makes it possible to have some special characters in the
identifiers of most of the constructs in Relational DDL and enables
querying on them. Since the tables are defined as Descriptor and stored
in DynamicMessage, it is imperative for table and column names to be a
valid symbol in protoBuf, which is limited to a-z, A-Z, 0-9 and _.
Hence, it support (slightly) richer character set, this PR mainly uses
surgical translations over names to make the RecordLayer Type system and
Protobuf types happy.

However, since the Relational Type and RecordLayer type themselves
honour the fact that the Relational name and RecordLayer name should be
the same, the translation spills to the topmost layer and is done quite
preemptively. To alleviate the issue, we might want to have a more
intrinsic understanding of "display" name and "underlying" name.

Another issue, comes from the fact that data flows through the execution
plan in protobuf Dynamic Messages, which makes it imperative for the
query aliases to be protobuf compliant as well!. We could potentially
break out of this by doing some schenanigans around what the recordLayer
type puts into the TypeRepository, but the work here takes a relatively
simpler approach of just making all aliases and user names
protobuf-friendly.

Note that, not all contructs and names are affecting. For example, index
names, schema template names, schema and database names should support
unicode charset out of the box. However, this PR does not really test
around that. I tried testing some part of it - while index names and
schema template names do work, databases naming is restrictive currently
and does not work. So, the scope of this PR is (or, should be)
everything else other than what is mentioned above.

From the implementation POV, the work takes a longer path of doing point
translations in many places rather than just one translation in
IdentifierVisitor to preemptively translate all identifiers. This is
owing to the following reasons:
- We would still want to leave some constructs from this translation -
those that do not require one. We could rather argue that translation
could be done even if not needed. However, this would have complicated
things around customer facing operations like `show`, `describe` items.
- Theoretically, we could very soon drift apart from using protobuf for
runtime data types flowing through plans. Once on it, the scope of
translation would reduce to barely to storage access operators -
supporting the fact that we should not do translations blindly.